cqrs-demo/consumer/requirements.txt
kafka-python==2.2.15
opensearch-py==3.0.0
